Move util to _util
