+++ src/talondoc/_util/progress_bar.py
