+++ src/talondoc/registry/entries/abc.py
@@ -10,7 +10,7 @@ from tree_sitter_talon import Node as Node
 from tree_sitter_talon import Point as Point
 from typing_extensions import Literal, TypeAlias, TypeVar, final
 
-from ...util.logging import getLogger
+from ..._util.logging import getLogger
 
 _LOGGER = getLogger(__name__)
